Improve verbose version output
