+++ cli/cli.go
@@ -34,6 +34,8 @@ import (
 	knfr "github.com/essentialkaos/ek/v12/knf/validators/regexp"
 	knfs "github.com/essentialkaos/ek/v12/knf/validators/system"
 
+	"github.com/essentialkaos/rep/cli/support"
+
 	"github.com/essentialkaos/rep/repo/index"
 )
 
@@ -241,7 +243,7 @@ func Init(gitRev string, gomod []byte) {
 		showAbout(gitRev)
 		return
 	case options.GetB(OPT_VERB_VER):
-		showVerboseAbout(gitRev, gomod)
+		support.ShowSupportInfo(APP, VER, gitRev, gomod)
 		return
 	case options.GetB(OPT_HELP) || len(args) == 0:
 		showUsage()
